Update
1. Step 1: Prerequisites -> Prepare your development environment
It should be a Ubuntu machine (for cross compiling), case the Seeeduino
Cloud run base on openWRT.

2. Step 3: Build and Run the sample.
The SDK Building environment is a Ubuntu machine, not on the Seeeduino
Cloud board. So the steps from "Install packages ..." to "Use scp
command ..." are all run on the Ubuntu machine.
The Steps "install ca-certificates" and "Change the files as a
executable file" are run on the Seeeduino Cloud board.

3. Add a scp command content on Step 3.1 -> "Use scp command ...", as
follows:
scp
~/openwrt/sdk/build_dir/target-mips_r2_uClibc-0.9.33.2/azure-iot-sdks-1/serializer/samples/iothub_client_sample_amqp/iothub_client_sample_amqp
root@Seeeduino_Cloud_IP_Address:/root
or
scp
~/openwrt/sdk/build_dir/target-mips_r2_uClibc-0.9.33.2/azure-iot-sdks-1/serializer/samples/iothub_client_sample_mqtt/iothub_client_sample_mqtt
root@Seeeduino_Cloud_IP_Address:/root